batch_ortho
Hi Saif,

I added the batch ortho function to the code. I also made two small modification to the correlate and batch_correlate sub_parser so that the output products go to the current directory.

I made additional edits to gcpOptimization.py so that each GGP file has the name of the images  (more practical when doing a batch ortho). 
I am also trying to get the file "CorregistrationError.png" (produced in geoTiePoints/misc.py) to have the image name included, but I could not manage yet. If you could take a look, that would be helpful.

Cheers,
Solene

Signed-off-by: soleneantoine1 <74420041+soleneantoine1@users.noreply.github.com>